split auth provider config
