+++ src/types/auth.d.ts
@@ -1,10 +1,9 @@
-import NextAuth, { type User } from "next-auth";
+import NextAuth from "next-auth";
 
 declare module "next-auth/jwt" {
 	interface JWT {
 		access_token: string;
 		access_token_expires?: number;
 		refresh_token: string;
-		user: User;
 	}
 }
